Silence an aligned access warning.
clang is a bit too strict about validating address alignments
when using packed structs.  The only member is a 32-bit value,
so the alignment was correct, but this approach is "correct"
and guaranteed to work even if the structure was not aligned,
at the cost of an extra temporary variable and 32-bit copy.